Inline overlayMemory accessors; move OverlayGroup taxonomy to constants
getOverlayVariable/setOverlayVariable were one-liners wrapping a plain
Record index -- inlined at all three call sites (overlayMemory.ts's
own activateOverlayGroup, MapPanel.tsx's two writes, Settings.tsx's
one read) rather than kept as indirection with no remaining logic of
their own.

OverlayGroup, OVERLAY_GROUP_SUMMARY_TYPES, and toOverlayGroup are pure
taxonomy with no store dependency -- moved from overlayMemory.ts
(utils) to constants/demography/summary.ts, alongside SummaryType/
isCoalitionUniverse/COALITION_UNIVERSES, the existing home for this
kind of demography-taxonomy definition. overlayGroupVariables stays in
overlayMemory.ts since it reads useDemographyStore.

districtr-v2-noq
