+++ app/src/app/components/sidebar/MapPanel.tsx
@@ -36,6 +36,7 @@ import {
   isCoalitionUniverse,
   CoalitionUniverse,
   SUMMARY_TYPES,
+  toOverlayGroup,
   type SummaryType,
 } from '@constants/demography/summary';
 import {COALITION_VARIABLE_BY_UNIVERSE, DemographyVariable} from '@constants/demography/coalition';
@@ -43,7 +44,7 @@ import {getCoalitionLabel, getSelectedCoalitionColumns} from '@/app/utils/demogr
 import {MAP_MODES} from '@constants/map/mode';
 import {NUMBER_FORMATS} from '@constants/demography/format';
 import {DEMOGRAPHIC_MODES} from '@constants/map/demographicMode';
-import {overlayMemory, setOverlayVariable, toOverlayGroup} from '@utils/demography/overlayMemory';
+import {overlayMemory} from '@utils/demography/overlayMemory';
 
 type MapPanelProps = {
   columnGroup: keyof typeof choroplethMapVariables;
@@ -176,7 +177,8 @@ export const MapPanel: React.FC<MapPanelProps> = ({columnGroup}) => {
     if (newMode) {
       overlayMemory.displayMode = newMode;
       const effectiveVariable = mapVariableConfig ? variable : currentVariableList[0]?.value;
-      if (effectiveVariable) setOverlayVariable(toOverlayGroup(columnGroup), effectiveVariable);
+      if (effectiveVariable)
+        overlayMemory.variables[toOverlayGroup(columnGroup)] = effectiveVariable;
     }
     if (!mapVariableConfig && currentVariableList.length) {
       setVariable(currentVariableList[0].value);
@@ -221,7 +223,7 @@ export const MapPanel: React.FC<MapPanelProps> = ({columnGroup}) => {
   const handleChangeVariable = (newVariable: DemographyVariable) => {
     setVariable(newVariable);
     // Remember the choice so the Visual settings overlay toggle restores it.
-    setOverlayVariable(toOverlayGroup(columnGroup), newVariable);
+    overlayMemory.variables[toOverlayGroup(columnGroup)] = newVariable;
   };
 
   const handleChangePercent = (usePercent: boolean) => {
