+++ app/src/app/constants/demography/summary.ts
@@ -20,3 +20,19 @@ export const TOTAL_COLUMN: Record<SummaryType, string | undefined> = {
   TOTPOP: 'total_pop_20',
   VOTERHISTORY: undefined,
 } as const;
+
+/**
+ * The two top-level choropleth overlay toggles. Coarser than SummaryType:
+ * the population overlay spans both the TOTPOP and VAP statistical
+ * universes as one merged control, not two.
+ */
+export type OverlayGroup = 'demography' | 'election';
+
+export const OVERLAY_GROUP_SUMMARY_TYPES: Record<OverlayGroup, SummaryType[]> = {
+  demography: [SUMMARY_TYPES.TOTPOP, SUMMARY_TYPES.VAP],
+  election: [SUMMARY_TYPES.VOTERHISTORY],
+};
+
+/** Which overlay group a statistical universe's variables belong to. */
+export const toOverlayGroup = (summaryType: SummaryType): OverlayGroup =>
+  summaryType === SUMMARY_TYPES.VOTERHISTORY ? 'election' : 'demography';
